Stopped it from updating oddly.
Hopefully added security ¯\_(ツ)_/¯

+++ UpdateMe.py
@@ -11,24 +11,28 @@ import sqlite3
 if __name__ == "__main__":
 	try:
 		database = jewsql.Database('jewsofhazard')
-		
+		user_values = database.get_user_values()
+		read_posts = database.get_post_values()
+		subreddits = list()
+		users = list()
+		checked_posts = list()
+		for row in user_values:
+			subreddits.append(row[1])
+			users.append(row[0])
+		for row in read_posts:
+			checked_posts.append(row[0])
+		print (users)
+		print (subreddits)
+		print (checked_posts)
 		r = praw.Reddit(user_agent='/u/JewsOfHazard made UpdateMe Script')
 		o = OAuth2Util.OAuth2Util(r)
+		notified_users = list()
+		file = open('notified_users.txt', 'r+')
+		for item in file:
+			notified_users.append(item)
 		while True:
 			try:
-				user_values = database.get_user_values()
-				read_posts = database.get_post_values()
-				subreddits = list()
-				users = list()
-				checked_posts = list()
-				for row in user_values:
-					subreddits.append(row[1])
-					users.append(row[0])
-				for row in read_posts:
-					checked_posts.append(row[0])
-				print (users)
-				print (subreddits)
-				print (checked_posts)
+
 				temp_checked_posts = list()
 				o.refresh()
 				for user in users:
@@ -52,9 +56,12 @@ if __name__ == "__main__":
 										print("A post has a character not supported.")
 										traceback.print_exc()
 						except praw.errors.Forbidden:
-							r.send_message(user, "I'm sorry, but I can't view posts from {}".format(sub), "For the time being, I have removed {0} from your list of followed subreddits. Please contact the moderators of {0} to allow me to view their subreddit.".format(sub))
-							database.delete_user_subreddit(user,sub)
-							print(sub, 'forbidden')
+							if user not in notified_users:
+								r.send_message(user, "I'm sorry, but I can't view posts from {}".format(sub), "For the time being, I have removed {0} from your list of followed subreddits. Please contact the moderators of {0} to allow me to view their subreddit.".format(sub))
+								database.delete_user_subreddit(user,sub)
+								notified_users.append(user)
+								file.write(user + '\n')
+								print(sub, 'forbidden')
 			except praw.errors.RateLimitExceeded:
 				print('Rate limit exceeded, sleeping 10 minutes.')
 				time.sleep(300)
